feat: implement add utility functions for testing and testing
