+++ jac/jaclang/langserve/tests/server_test/test_lang_serve.py
@@ -1,186 +1,171 @@
+"""Test for Jac language server[VSCE] features"""
 import os
-import asyncio
 import pytest
-import contextlib
 
 from lsprotocol.types import (
     DidOpenTextDocumentParams,
     TextDocumentItem,
     DidSaveTextDocumentParams,
     DidChangeTextDocumentParams,
-    TextDocumentContentChangeEvent,
     DocumentFormattingParams,
-    CreateFilesParams,
-    FileCreate,
-    Position,
+    TextEdit,
     VersionedTextDocumentIdentifier,
     TextDocumentIdentifier,
 )
 from jaclang.langserve.engine import JacLangServer
-from jaclang.langserve.server import did_open, did_save, did_change, formatting, did_create_files
+from jaclang.langserve.server import (
+    did_open,
+    did_save,
+    did_change,
+    formatting,
+)
+from jaclang.langserve.tests.server_test.utils import get_code, patch_file, get_jac_file_path
 from jaclang.vendor.pygls.uris import from_fs_path
 from jaclang.vendor.pygls.workspace import Workspace
 
+JAC_FILE = get_jac_file_path()
 
-# Path to a sample Jac file for testing
-JAC_FILE = os.path.abspath(
-    os.path.join(os.path.dirname(__file__), "../../../../examples/manual_code/circle.jac")
-)
+class TestLangServe:
+    @pytest.mark.asyncio
+    async def test_did_open_and_semantic_tokens(self):
+        """Test opening a Jac file and retrieving semantic tokens."""
+        ls: JacLangServer = JacLangServer()
+        uri = from_fs_path(JAC_FILE)
+        ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
 
-@pytest.mark.asyncio
-async def test_did_open_and_semantic_tokens():
-    # Create a fresh instance of the language server
-    ls :JacLangServer= JacLangServer()
-
-    uri = from_fs_path(JAC_FILE)
-
-    # Set up workspace so the server can track the document
-    ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
-
-    with open(JAC_FILE, "r") as f:
-        code = f.read()
-
-    # Create didOpen parameters
-    params = DidOpenTextDocumentParams(
-        text_document=TextDocumentItem(
-            uri=uri,
-            language_id="jac",
-            version=1,
-            text=code,
-        )
-    )
-
-    # Directly call the `did_open` feature handler
-    await did_open(ls, params)
-
-    # Validate diagnostics
-    diagnostics = ls.diagnostics.get(uri, [])
-    print("Diagnostics:", diagnostics)
-    assert isinstance(diagnostics, list)
-    assert len(diagnostics) == 0
-
-    # Validate semantic tokens
-    sem_tokens = ls.get_semantic_tokens(uri)
-    print("Semantic Tokens:", sem_tokens)
-    assert hasattr(sem_tokens, "data")
-    assert isinstance(sem_tokens.data, list)
-
-@contextlib.contextmanager
-def patch_file(filepath, new_content):
-    """Temporarily replace file content for testing."""
-    with open(filepath, "r") as f:
-        original = f.read()
-    try:
-        with open(filepath, "w") as f:
-            f.write(new_content)
-        yield
-    finally:
-        with open(filepath, "w") as f:
-            f.write(original)
-
-@pytest.mark.asyncio
-async def test_did_open_and_simple_syntax_error():
-    ls = JacLangServer()
-    uri = from_fs_path(JAC_FILE)
-    ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
-
-    from jaclang.langserve.tests.server_test.code_test import get_code
-    broken_code = get_code("error")
-
-    with patch_file(JAC_FILE, broken_code):
+        code = get_code("error")
         params = DidOpenTextDocumentParams(
             text_document=TextDocumentItem(
                 uri=uri,
                 language_id="jac",
                 version=1,
-                text=broken_code,
+                text=code,
             )
         )
         await did_open(ls, params)
+
         diagnostics = ls.diagnostics.get(uri, [])
-        print("Diagnostics:", diagnostics)
         assert isinstance(diagnostics, list)
-        assert len(diagnostics) == 1
+        assert len(diagnostics) == 0
 
         sem_tokens = ls.get_semantic_tokens(uri)
-        print("Semantic Tokens:", sem_tokens)
         assert hasattr(sem_tokens, "data")
         assert isinstance(sem_tokens.data, list)
+        assert len(sem_tokens.data) == 300
 
-@pytest.mark.asyncio
-async def test_did_save():
-    ls = JacLangServer()
-    uri = from_fs_path(JAC_FILE)
-    ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
-    with open(JAC_FILE, "r") as f:
-        code = f.read()
-    # Open document first
-    await did_open(ls, DidOpenTextDocumentParams(
-        text_document=TextDocumentItem(
-            uri=uri,
-            language_id="jac",
-            version=1,
-            text=code,
+    @pytest.mark.asyncio
+    async def test_did_open_and_simple_syntax_error(self):
+        """Test diagnostics for a Jac file with a syntax error."""
+        ls = JacLangServer()
+        uri = from_fs_path(JAC_FILE)
+        ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
+
+        broken_code = get_code("error")
+
+        with patch_file(JAC_FILE, broken_code):
+            params = DidOpenTextDocumentParams(
+                text_document=TextDocumentItem(
+                    uri=uri,
+                    language_id="jac",
+                    version=1,
+                    text=broken_code,
+                )
+            )
+            await did_open(ls, params)
+            diagnostics = ls.diagnostics.get(uri, [])
+            assert isinstance(diagnostics, list)
+            assert len(diagnostics) == 1
+
+            sem_tokens = ls.get_semantic_tokens(uri)
+            assert hasattr(sem_tokens, "data")
+            assert isinstance(sem_tokens.data, list)
+
+    @pytest.mark.asyncio
+    async def test_did_save(self):
+        """Test saving a Jac file triggers diagnostics."""
+        ls = JacLangServer()
+        uri = from_fs_path(JAC_FILE)
+        ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
+
+        code = get_code("")
+
+        await did_open(ls, DidOpenTextDocumentParams(
+            text_document=TextDocumentItem(
+                uri=uri,
+                language_id="jac",
+                version=1,
+                text=code,
+            )
+        ))
+
+        params = DidSaveTextDocumentParams(
+            text_document=TextDocumentItem(
+                uri=uri,
+                language_id="jac",
+                version=2,
+                text=code,
+            )
         )
-    ))
-    # Save event
-    params = DidSaveTextDocumentParams(
-        text_document=TextDocumentItem(
-            uri=uri,
-            language_id="jac",
-            version=2,
-            text=code,
+
+        await did_save(ls, params)
+        diagnostics = ls.diagnostics.get(uri, [])
+        assert isinstance(diagnostics, list)
+        assert len(diagnostics) == 0
+
+        broken_code = get_code("error")
+        with patch_file(JAC_FILE, broken_code):
+            params = DidOpenTextDocumentParams(
+                text_document=TextDocumentItem(
+                    uri=uri,
+                    language_id="jac",
+                    version=1,
+                    text=broken_code,
+                )
+            )
+
+            await did_open(ls, params)
+            diagnostics = ls.diagnostics.get(uri, [])
+            assert isinstance(diagnostics, list)
+            assert len(diagnostics) == 1
+
+    @pytest.mark.asyncio
+    async def test_did_change(self):
+        """Test changing a Jac file triggers diagnostics."""
+        ls = JacLangServer()
+        uri = from_fs_path(JAC_FILE)
+        ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
+        with open(JAC_FILE, "r") as f:
+            code = f.read()
+
+        await did_open(ls, DidOpenTextDocumentParams(
+            text_document=TextDocumentItem(
+                uri=uri,
+                language_id="jac",
+                version=1,
+                text=code,
+            )
+        ))
+
+        params = DidChangeTextDocumentParams(
+            text_document=VersionedTextDocumentIdentifier(uri=uri, version=2),
+            content_changes=[{"text": "\n"+ code }]
         )
-    )
-    await did_save(ls, params)
-    await asyncio.sleep(1)
-    diagnostics = ls.diagnostics.get(uri, [])
-    assert isinstance(diagnostics, list)
+        await did_change(ls, params)
+        diagnostics = ls.diagnostics.get(uri, [])
+        uri = from_fs_path(JAC_FILE)
+        assert isinstance(diagnostics, list)
+        assert len(diagnostics) == 0
 
-@pytest.mark.asyncio
-async def test_did_change():
-    ls = JacLangServer()
-    uri = from_fs_path(JAC_FILE)
-    ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
-    with open(JAC_FILE, "r") as f:
-        code = f.read()
-    # Open document first
-    await did_open(ls, DidOpenTextDocumentParams(
-        text_document=TextDocumentItem(
-            uri=uri,
-            language_id="jac",
-            version=1,
-            text=code,
+    def test_vsce_formatting(self):
+        """Test formatting a Jac file returns edits."""
+        ls = JacLangServer()
+        uri = from_fs_path(JAC_FILE)
+        ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
+        params = DocumentFormattingParams(
+            text_document=TextDocumentIdentifier(uri=uri),
+            options={"tabSize": 4, "insertSpaces": True}
         )
-    ))
-    # Change event
-    params = DidChangeTextDocumentParams(
-        text_document=VersionedTextDocumentIdentifier(uri=uri, version=2),
-        content_changes=[{"text": code + "\n"}]
-    )
-    await did_change(ls, params)
-    await asyncio.sleep(1)
-    diagnostics = ls.diagnostics.get(uri, [])
-    assert isinstance(diagnostics, list)
-
-def test_formatting():
-    ls = JacLangServer()
-    uri = from_fs_path(JAC_FILE)
-    ls.lsp._workspace = Workspace(os.path.dirname(JAC_FILE), ls)
-    # Formatting params
-    params = DocumentFormattingParams(
-        text_document=TextDocumentIdentifier(uri=uri),
-        options={"tabSize": 4, "insertSpaces": True}
-    )
-    edits = formatting(ls, params)
-    assert isinstance(edits, list)
-
-def test_did_create_files():
-    ls = JacLangServer()
-    uri = from_fs_path(JAC_FILE)
-    params = CreateFilesParams(
-        files=[FileCreate(uri=uri)]
-    )
-    # Should not raise
-    did_create_files(ls, params)
-    print(ls.diagnostics)
+        edits = formatting(ls, params)
+        assert isinstance(edits, list)
+        assert isinstance(edits[0], TextEdit)
+        assert len(edits[0].new_text) > 100 # it is a random number to check if the text is changed
